Adding game over screen
Comfirming quit on the main level 'should' go to the game over screen, but for some reason right now it is going back to the starting screen, which may still work.
Restart button 'should' go back to the main level, but this is untested as we can't go to this game over screen yet.